Stable changes using APC to inject Sanctum DLL into target process.
Currently only targeting malware.exe and notepad.exe; targeting wider applications is causing some bugs with regard to the DLL not being correctly signed etc, so that likely needs an environment change and some further testing.. but this works, and in stable!

There is a bug with Ghost Hunting where it is not picking up on some NtOpenProcess calls.. very annoying, and needs investigating.